CDSK-568 - move add_css_classes_to_results and tests to separate files
